Change method of RL
instead of passing a previous net as the input to each time-step. Pass the previous decided action. The first input is always convolution which is added straight to the action list.
Add checking to whether we have already tested a net before we do it again and waste time. Also changed the reward for impossible nets to 0.